Broke space type components into own files
