+++ src/components/Space/types.ts
