--- src/models/space/index.ts
@@ -1,5 +1,0 @@
-export interface SpaceState {
-  path: Boolean;
-  type: string;
-  visited: Boolean;
-}
